feat(auth): configuration initiale de Supabase et authentification
- Configuration de la connexion à Supabase
- Mise en place des variables d'environnement
- Création des tables
- Test de l'authentification et création d'utilisateur
- Nettoyage des routes de test
- Ajout des règles de documentation dans .cursorrules
- Mise en place de la structure de base du projet (routes, controllers, models, middlewares)
